Refactor shared code to module
