+++ test/lib/tasks/developer_digest_task_test.rb
@@ -1,11 +1,12 @@
 require "test_helper"
 
 class DeveloperDigestTaskTest < ActionMailer::TestCase
+  include DevelopersHelper
   include RakeTaskHelper
 
   test "sends appropriate emails for developer digests" do
     load_rake_tasks_once
-    Developer.create!(valid_developer_attributes.merge(created_at: 1.day.ago))
+    Developer.create!(developer_attributes.merge(created_at: 1.day.ago))
 
     assert_emails 1 do
       Rake::Task["developer_digest:daily"].invoke
